Prepare for Gulp Task Split
+ Installs `require-dir` module.
+ Moves `gulpfile.js` to `gulp/tasks/combined.js`.
+ Adds one liner `gulpfile.js`.
+ Replace `__dirname` with `backstopPath` in `gulp/tasks/combined.js`.
