+++ package.json
@@ -12,13 +12,14 @@
   },
   "dependencies": {
     "bunyan": "~0.22.3",
-    "express": "^3.8.1",
-    "http-proxy": "~1.1.4",
-    "underscore": "~1.6.0",
-    "gulp": "^3.8.8",
     "del": "^0.1.3",
+    "express": "^3.8.1",
+    "fs-extra": "^0.12.0",
+    "gulp": "^3.8.8",
     "gulp-open": "^0.3.0",
     "gulp-rename": "^1.2.0",
-    "fs-extra": "^0.12.0"
+    "http-proxy": "~1.1.4",
+    "require-dir": "^0.1.0",
+    "underscore": "~1.6.0"
   }
 }
